% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{alter_cols}
\alias{alter_cols}
\title{Function that uniformly applies the same change to many columns. Intended to do things like 
converting character columns to numeric ones.}
\usage{
alter_cols(d, col_finder, alterFunc)
}
\arguments{
\item{d}{Input data frame needing column alteration.}

\item{col_finder}{Take the whole data frame (d here) and return a vector of numeric indices of columns to alter.}

\item{alterFunc}{Takes a single vector, returns a single vector that is an altered copy.}
}
\value{
A copy of the input data frame d with the given columns altered.
}
\description{
Function that uniformly applies the same change to many columns. Intended to do things like 
converting character columns to numeric ones.
}
