% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{col_norm_null}
\alias{col_norm_null}
\title{col_norm_null}
\usage{
col_norm_null(dd, normCols, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame}

\item{normCols}{dummy variable for column to normalize with. This whole function does nothing.}

\item{rowInds}{dummy variable}

\item{colInds}{dummy variable}
}
\value{
returns the input data frame, exactly the way it was.
}
\description{
col_norm_null doesn't do any normalizing, but represents. Use this if you need a default normalizing arg
that does nothing.
}

