% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cols.R
\name{col_norm_to_median}
\alias{col_norm_to_median}
\title{col_norm_to_median normalizes all the frame-columns by the median of a number of them..}
\usage{
col_norm_to_median(dd, normRows, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normRows}{A vector of indices that pick rows. Medians are calculated from these in column-wise manner, and
they are used to normalize the corresponding column. May be a function like 'span'}

\item{rowInds}{A span of row indices over which to operate. These should be the data rows, and this is
just so that you can have other ignored rows be in there. May be a function like 'span'.}

\item{colInds}{A span of indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there. May be a function like 'span'.}
}
\value{
returns the input data frame, but with the specified data normalized.
}
\description{
col_norm_to_median normalizes all the frame-columns by the median of a number of them..
}
