% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{col_norm_to_sum}
\alias{col_norm_to_sum}
\title{col_norm_to_sum normalizes all the frame-columns by the sum of a number of them..}
\usage{
col_norm_to_sum(dd, normCols, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normCols}{A vector of indices that pick columns. Sums (for normalization) are calculated from these in row-wise manner, and
they are used to normalize. May be a function like 'span'.}

\item{inds}{A span of indices over which to operate. Only these columns will have their entries normalized. Useful for ignoring
id columns etc. May be a function like 'span'.}
}
\value{
returns the input data frame, but with the data cols normalized.
}
\description{
col_norm_to_sum normalizes all the frame-columns by the sum of a number of them..
}

