% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{generate_col}
\alias{generate_col}
\title{Function that somehow generates a new column called "new_col_name" from the old column, called
"col_name". splitFunc is a function that accepts a vector, and returns a new vector. This function
is intended to do things like split sample names from the default values that are multipart and
full of "_"s.}
\usage{
generate_col(d, col_name, new_col_name, splitFunc)
}
\arguments{
\item{d}{Input data frame needing new_col_name from col_name}

\item{col_name}{The name of the column to use in generating the new one.}

\item{new_col_name}{What are we going to name the new column.}

\item{splitFunc}{Takes a single vector, returns a single vector, and contains the column creation logic.}
}
\value{
A copy of the input dataframe d, with the new column appended
}
\description{
Function that somehow generates a new column called "new_col_name" from the old column, called
"col_name". splitFunc is a function that accepts a vector, and returns a new vector. This function
is intended to do things like split sample names from the default values that are multipart and
full of "_"s.
}

