% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{keep_rows}
\alias{keep_rows}
\title{keep_rows}
\usage{
keep_rows(d, funcList)
}
\arguments{
\item{d}{Input data frame}

\item{funcList}{a list of functions. Each takes a whole dataframe, and returns a vector of indices detailing which rows to keep.}
}
\value{
The input data frame \code{d} but containing only the rows we want to keep
}
\description{
Function keeps only the specified rows from input data frame d, returning a new
data frame containing only these rows.
}

