% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_frames.R
\name{massShift_filter}
\alias{massShift_filter}
\title{massShift_filter removes peaks that correspond to specific mass shifts.}
\usage{
massShift_filter(d, mzColName = "MZ", timeColName = "Time",
  shiftList1 = to.elim.1, shiftList2 = to.elim.2, rt_window_width = 0.05)
}
\arguments{
\item{mzColName}{This is the name of the MZ column in the sieve data to be filtered.}

\item{timeColName}{This is the retention time column name in the sieve data to be filtered.}

\item{shiftList1}{Primary list of mass shifts. These should correspond to things like adducts and fragments
that additionally have other shifts (like isotopes), provided in shiftList2, calculated.}

\item{shiftList2}{Secondary list of mass shifts. This should correspond to things like isotopes, that need to be
calculated for each peak in the dataset, and also for each added fragment peak considered from shiftList1.}

\item{rt_window_width}{Peaks are only considered as fragments/isotopes if they're within a particular chromatographic
retention time window of the parent peak. How wide is that window? Default: 0.05}

\item{mass_acc}{The mass accuracy of the instrument, when we're within the rt window, looking for the mass of a fragment peak,
we consider everything within a +/- mass_acc window from the target value.}
}
\description{
massShift_filter removes peaks that correspond to specific mass shifts.
}

