% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_null}
\alias{norm_null}
\title{norm_null}
\usage{
norm_null(dd, normCols, inds = NULL)
}
\arguments{
\item{dd}{input data frame}

\item{normCols}{dummy variable for column to normalize with. This whole function does nothing.}

\item{inds}{dummy variable}
}
\value{
returns the input data frame, exactly the way it was.
}
\description{
norm_null doesn't do any normalizing, but represents. Use this if you need a default normalizing arg
that does nothing.
}

