% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_to_frame}
\alias{norm_to_frame}
\title{norm_to_frame normalizes all the frame-columns by one of them.}
\usage{
norm_to_frame(dd, normCol, inds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normCol}{either a string column name, or a numeric column index}

\item{inds}{A span of indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there. This can be NULL, in which case all columns
  are normalized, or it can be a vector of integer indices, or it can be a function that gets called with dd
  and which returns a vector of ints (see function 'span' in this package)}
}
\description{
Set this to the first frame to normalize by the most intense frame, or set this to
the internal standard frame.
}

