% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{norm_to_median}
\alias{norm_to_median}
\title{norm_to_median normalizes all the frame-columns by the median of a number of them..}
\usage{
norm_to_median(dd, normCols, inds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normCols}{A vector of indices that pick columns. Medians are calculated from these in row-wise manner, and
they are used to normalize.}

\item{inds}{A span of indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there.}
}
\value{
returns the input data frame, but with the data cols normalized.
}
\description{
norm_to_median normalizes all the frame-columns by the median of a number of them..
}

