% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{p}
\alias{p}
\title{P is the pipelining function of this module. It decorates functions such that their only argument
becomes the data frame in the pipeline.
Intended for use with magrittr pipelines.}
\usage{
p(f)
}
\description{
P is the pipelining function of this module. It decorates functions such that their only argument
becomes the data frame in the pipeline.
Intended for use with magrittr pipelines.
}
