% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{print_dim}
\alias{print_dim}
\title{print_dim}
\usage{
print_dim(d, tagStr)
}
\arguments{
\item{d}{The input data frame for dim printing.}

\item{tagStr}{A string to provide some context}
}
\value{
Prints to the screen... returns the data frame d for further pipelining.
}
\description{
print_dim simply prints the dim of input matrix d using "print"
}
