% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{read_data_file}
\alias{read_data_file}
\title{Function reads in a csv file, and prepares it so that it has a proper header line
(1st line), and a consistent body afterwards, with no trailing empty rows or columns.}
\usage{
read_data_file(fname, body = NULL)
}
\arguments{
\item{fname}{This is the filename to read in}

\item{body}{Need something other than the read.csv line described below? Pass it in here.}
}
\value{
a data frame, with something sensical in names(...)
}
\description{
Function reads in a csv file, and prepares it so that it has a proper header line
(1st line), and a consistent body afterwards, with no trailing empty rows or columns.
}

