% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remCorlowIntensity.R
\name{remCor.lowintensity}
\alias{remCor.lowintensity}
\title{remCor.lowintensity}
\usage{
remCor.lowintensity(d, cutoff = 0.8, measure = median)
}
\arguments{
\item{d}{Input data frame}

\item{cutoff}{Correlation threshold at which to do removal}

\item{measure}{Column measure used to determine column to remove}
}
\value{
a vector of column indices to keep.
}
\description{
Filter the columns of data frame d such that the less median-intense (or other user
specifiable column, correlated at a value higher than cutoff, is removed.
}

