% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{remove_cols}
\alias{remove_cols}
\title{Function removes columns from input data frame d, returning a data frame without them}
\usage{
remove_cols(d, funcList, inds = NULL)
}
\arguments{
\item{d}{Input data frame needing columns removed}

\item{funcList}{A list of functions that each yield column indices to be removed. Called with dataframe as input. Joined with union.}

\item{inds}{A vector of indexes to consider for removal. Useful for guaranteeing that id info won't be removed etc. The final list of columns to remove is intersected with this list. This can also be a function that is called with the data frame d, and returns a vector of indices (such as the 'span' function above).}
}
\value{
The input data frame \code{d} but missing the unwanted columns
}
\description{
Function removes columns from input data frame d, returning a data frame without them
}

