% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{remove_rows}
\alias{remove_rows}
\title{Function removes rows from input data frame d, returning a data frame without them}
\usage{
remove_rows(d, funcList, inds = NULL)
}
\arguments{
\item{d}{Input data frame needing rows removed}

\item{funcList}{a list of functions. Each takes a whole dataframe, and returns a vector of indices detailing which rows to remove.}

\item{inds}{A vector of indexes to consider for removal. Useful for guaranteeing that id info won't be removed etc. The final list of rows to remove is intersected with this list. This can also be a function that is called with the data frame d, and returns a vector of indices (such as the 'span' function above).}
}
\value{
The input data frame \code{d} but missing the unwanted rows
}
\description{
Function removes rows from input data frame d, returning a data frame without them
}

