% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_rows.R
\name{row_norm_to_sum}
\alias{row_norm_to_sum}
\title{row_norm_to_sum normalizes all the frame-columns by the sum of a number of them..}
\usage{
row_norm_to_sum(dd, normRows, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normRows}{A vector of indices that pick rows. Sums are calculated from these in column-wise manner, and
they are used to normalize. May be a function like 'span'}

\item{rowInds}{A span of row indices over which to operate. These should be the data rows, and this is
just so that you can have other ignored rows be in there. May be a function like 'span'.}

\item{colInds}{A span of indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there. May be a function like 'span'.}
}
\value{
returns the input data frame, but with the specified data normalized.
}
\description{
row_norm_to_sum normalizes all the frame-columns by the sum of a number of them..
}

