% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{span}
\alias{span}
\title{Function that describes a range in a given data frame context. Intended for use within a
pipeline, where there's no explicit data frame to measure, like for the inds argument of
remove_cols and remove_rows.}
\usage{
span(type, fromInd, toInd)
}
\arguments{
\item{type}{Either Is this a range of 'row' or 'col'}

\item{fromInd}{Either a numeric, or 'start' or 'end'}

\item{toInd}{Either a numeric, or 'start' or 'end'}
}
\value{
function(d), called with a data frame, that returns a numeric vector.
}
\description{
Function that describes a range in a given data frame context. Intended for use within a
pipeline, where there's no explicit data frame to measure, like for the inds argument of
remove_cols and remove_rows.
}

