% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import-data.R
\name{transpose}
\alias{transpose}
\title{Function transposes the dataframe. You tell it which columns contain row id information, and
which contain row measurement information, so that it can put together the transposed dataset correctly.
The "names" of the data frame become the row.names after the transpose. This is made into an additional column
called sample_id.}
\usage{
transpose(d, id_cols, id_col_name, name_cols = NULL)
}
\arguments{
\item{d}{Input data frame}

\item{id_cols}{numeric indices of columns containing row id information}

\item{id_col_name}{If we don't supply a name_cols function, what do we want to call the auto-generated id column?}

\item{name_cols}{unary function that takes the sub-matrix of (formerly row, after transpose column) information, and creates a vector
to be used for sample names out of it.}
}
\value{
transposed data frame, with the sample_ids in the first colmns (used to be original data frame names), and the frame/variable
        id information in the names(...) attribute.
}
\description{
Function transposes the dataframe. You tell it which columns contain row id information, and
which contain row measurement information, so that it can put together the transposed dataset correctly.
The "names" of the data frame become the row.names after the transpose. This is made into an additional column
called sample_id.
}

