% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_frames.R
\name{vasant_filter}
\alias{vasant_filter}
\title{vasant_filter runs Vasant's vectorized correlation filter on data frame d.}
\usage{
vasant_filter(d, corr_cutoff = 0.8, measure = median, inds = NULL)
}
\arguments{
\item{d}{data frame (already transposed) to filter.}

\item{corr_cutoff}{Correlation value above which to remove the less median-intense column. DEFAULT:0.8}

\item{inds}{Vector of inds over which to operate}
}
\value{
Copy of data frame d, but missing the columns (optionally specified by inds) that don't make the cutoff
}
\description{
vasant_filter runs Vasant's vectorized correlation filter on data frame d.
}
