% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accessories.R
\name{writeOut}
\alias{writeOut}
\title{writeOut writes dataframe dd out to a csv file named n.}
\usage{
writeOut(dd, n)
}
\arguments{
\item{dd}{the input data frame}

\item{n}{the name of the output csv file (include extension)}
}
\value{
This returns the data frame for further chaining with magrittr
}
\description{
writeOut writes dataframe dd out to a csv file named n.
}
