% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cols.R
\name{col_norm_to_median}
\alias{col_norm_to_median}
\title{col_norm_to_median normalizes all the frame-columns by the median of a number of them..}
\usage{
col_norm_to_median(dd, normCols, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normCols}{A vector of indices that pick columns. Medians are calculated from these in row-wise manner, and
they are used to normalize. May be a function like 'span'}

\item{rowInds}{A span of row indices over which to operate. These should be the data rows, and this is
just so that you can have other ignored rows be in there. This can be NULL, in which case all rows
  are normalized, or it can be a vector of integer indices, or it can be a function that gets called with dd
  and which returns a vector of ints (see function 'span' in this package)}

\item{colInds}{A span of column indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there. This can be NULL, in which case all columns
  are normalized, or it can be a vector of integer indices, or it can be a function that gets called with dd
  and which returns a vector of ints (see function 'span' in this package)}
}
\value{
returns the input data frame, but with the data cols normalized.
}
\description{
col_norm_to_median normalizes all the frame-columns by the median of a number of them..
}
