% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{col_norm_to_median}
\alias{col_norm_to_median}
\title{col_norm_to_median normalizes all the frame-columns by the median of a number of them..}
\usage{
col_norm_to_median(dd, normCols, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normCols}{A vector of indices that pick columns. Medians are calculated from these in row-wise manner, and
they are used to normalize. May be a function like 'span'}

\item{inds}{A span of indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there. May be a function like 'span'.}
}
\value{
returns the input data frame, but with the data cols normalized.
}
\description{
col_norm_to_median normalizes all the frame-columns by the median of a number of them..
}

