% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_cols.R
\name{col_norm_to_vector}
\alias{col_norm_to_vector}
\title{col_norm_to_vector normalized all the frame-columns by a supplied vector, same length as columns.
Pre-compute some other vector by which to normalize (like from a different but identically dimensioned matrix),
and use it to normalize the columns  of the input matrix dd.}
\usage{
col_norm_to_vector(dd, normVector, rowInds = NULL, colInds = NULL)
}
\arguments{
\item{dd}{input data frame to normalize}

\item{normVector}{This is a vector that will properly normalize the columns of this data frame, BEFORE
any non-processed rows are removed from it. (i.e. it has to have exactly the same dimension as
the pre-processed data frame it's going to normalize.)}

\item{rowInds}{A span of row indices over which to operate. These should be the data rows, and this is
just so that you can have other ignored rows be in there. This can be NULL, in which case all rows
  are normalized, or it can be a vector of integer indices, or it can be a function that gets called with dd
  and which returns a vector of ints (see function 'span' in this package)}

\item{colInds}{A span of column indices over which to operate. These should be the data columns, and this is
just so that you can have other ignored columns be in there. This can be NULL, in which case all columns
  are normalized, or it can be a vector of integer indices, or it can be a function that gets called with dd
  and which returns a vector of ints (see function 'span' in this package)}
}
\value{
return the input data frame, but with the specified data columns normalized by the input column vector.
}
\description{
col_norm_to_vector normalized all the frame-columns by a supplied vector, same length as columns.
Pre-compute some other vector by which to normalize (like from a different but identically dimensioned matrix),
and use it to normalize the columns  of the input matrix dd.
}

